+++ nix/checks/openclaw-package-contents.nix
@@ -3,7 +3,6 @@
   stdenv,
   nodejs_22,
   openclawGateway,
-  requireAgentWorkspaceTemplates ? true,
 }:
 
 stdenv.mkDerivation {
@@ -16,9 +15,6 @@ stdenv.mkDerivation {
 
   env = {
     OPENCLAW_GATEWAY = openclawGateway;
-  }
-  // lib.optionalAttrs (!requireAgentWorkspaceTemplates) {
-    OPENCLAW_REQUIRE_AGENT_WORKSPACE_TEMPLATES = "0";
   };
 
   doCheck = true;